ML/requirements.txt
# 🔧 Core
flask
flask-cors
pyyaml
tqdm
Pillow
scikit-learn

# ⚙️ Torch + Faiss
torch==1.11.0
torchvision==0.12.0
faiss-cpu

# 📷 Computer Vision
opencv-python==4.1.2.30
mediapipe
ultralytics
albumentations==0.4.3
kornia==0.6

# 🤗 NLP + CLIP
transformers==4.19.2

# 🧠 Face Analysis (Gender, Age, etc.)
insightface
onnxruntime

# 📦 Core + Utilities
numpy==1.19.2
diffusers
bezier
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
torchmetrics==0.6.0

# Editable GitHub dependencies
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
-e git+https://github.com/openai/CLIP.git@main#egg=clip
